bash: fix "getcwd: cannot access parent directories: Math result..."
In get_working_directory(), the getcwd call in the __ANDROID__
workaround case was obviously wrong, having sizeof(char *) as a length
argument instead of the length of the malloced string itself. Also, no
check for malloc failure was there.

With this fix, tested on maguro, using bash with the usual prompt
containing pwd references, as well as the pwd bash builtin itself,
no longer spew these annoying messages mentioned in the title.

Also made a small change to sh_invalidnum() to fix compile warnings.

Change-Id: I9546e07731a251f80e410935619ddcf0873694af